add nonclearing nonground segmented cloud type
Add a third (optional) segmented cloud type, nonclearing, nonground.
This type is useful to add points to the octomap as obstacles that
are not raytraced. This could happen for logical obstacles (such as
lack of sensed navigable floor) that should appear in a particular
place but do not represent a clear ray of space.