change dynamic reconfigure limits for min/max z
Remove the limits on min/max z and set the defaults to a very big
limits.  This allows a point cloud to publish points very far away to
represent a ray that should be cleared but not marked. The base
position bounds can be used to limit the ray tracing to only the
space that should be tracked.