add configurable spatial limits to updates
Add the ability to configure the spatial limits for updates to be
tighter than the limits for the whole map. This allows the map to
track a large space, but only update a local area. This is useful for
building a global map while limiting how much cpu time is spent ray
tracing to a local space.