OcTreeStampedWithExpiry: add option to only update expiry
Add an option to only update the expiry without deleting expired nodes.
This is helpful when publishing debug markers so the expiration can be
updated prior to publishing without altering the structure of the tree.
It is also more efficient than having to delete the nodes at the publish
rate.