segment clouds: pre-voxel-filter input clouds
Only use one or two voxels for any location during an update cycle.
The only time two voxels is used if the first point encountered in
that voxel is free, and later an occupied point is found in that
voxel. Pre-filtering greatly improves performance in the case where
points are within the clipping region, as none of the clipping vector
code needs to be run on every input point. Pre-filtering also
improves performance when combining two types of sensors with largely
overlapping fields of view. This might happen when using a structured
light camera and stereo camera over the same space.