sensor update: truncate floor along ray
Do not simply truncate z values for rays out-side of the
update bounds. Instead, move the ray as before, and mark the
end point if the ray was below the floor and is now at the floor
(but not above, in the case that more than one bound was violated).

This is much more accurate, as cliff like obstacles will appear at
the point the floor was not observable instead of at the perpendicular
projection.