publish sequence numbers on update sub-headers
These sequence numbers can be used to ensure no update message
is lost. The sequence number from ROS CPP can not be read by the
publisher, so there is no way to communicate on the first
full map what the next sequence number should be. This means
if the first update is lost, there is no way to know.