add layered updates, downSample each layer to make tree
Add layered updates to the sensor update. Create a downSample method
for each implementation which creates a down-sampled version of the
current layer. This creates a tree structure that can be used to very
efficiently apply the update to the tree by walking the tree and the
update only once instead of calling 'update' on every key.