do not enable octomap change detection
The octomap library was updated to no longer require change detection
to be enabled to call the value change callback on a change. Since we
use the value change callback to track changes in an octomap, there
is no benefit to tracking them additionally in the octomap itself.
Also, we were not resetting the change detection directly in the
octomap, leaking memory every update cycle.