make OcTreeStampedWithExpiry work for color octomaps
Templatize OcTreeStampedWithExpiry to take the base node type for its
occupancy octomaps (so either OcTreeNode or ColorOcTreeNode can be
used as the base node type). Set the appropriate base node type in
types.h for color vs. non-color octomaps. Move all the common code
between the color and plain octomap servers into a new
"octomap_server_common" library to avoid including the sensor update
code twice. Provide storage space for the two different types of
octomap nodes in the common library.

This change was required when rebasing on to 0.6.5 as the color server
is now a separate server that is always compiled.