OcTreeStampedWithExpiry: publish color octrees
After adding support for color octrees, the old OcTree type was still
being used to publish the tree. This was due to the following commit:

commit 5e4d85298b7e8d1119fe0a57abcb61400dd833d5
Author: jatravis-nj <jacob_travis@jabil.com>
Date:   Thu May 16 10:17:41 2019 -0400

    Temp change to allow rviz vizualization

Document why this commit was made (because the octomap_rviz_plugin
doesn't know anything about OcTreeStampedWithExpiry), and use
"ColorOcTree" if we are extending ColorOcTreeNode, otherwise use
"OcTree".