actually define static member initializer
The static member initializer was only being declared in the .cpp
file and not actually defined.

This error persisted for a while because the optimzer was hiding this
error as it would optimize the whole class away since the call to
register the tree with the abstract OcTree class was commented out a
while ago for visualization purposes.