optimizations: insertFreeRay, insertFreeCells
Optimize insertFreeRay and insertFreeCells.

Have insertFreeRay check that the end of the ray is in bounds
and change the check for bounds to an assert in the ray-tracing loop.
This is a very hot path for large point cloud sensors and the
extra test and branch add a lot of unnecessary overhead.

Have insertFreeCells only work in level 0 sensor update maps.
This is fine as it is only ever called on level 0 maps. The
new implementation will return false if called on a non-level 0 map.
Then, the inner calculateIndex can be made calculateIndexLevel0
which is inline static version with no branches. This is another
very hot path in the code.

Also, through emperical testing, it is found to be more effecient in
insertFreeCells for SensorKeyMapArrayImpl to use a ternary operator
always setting the current cell back to what it was when it was
not unknown instead of conditoinally setting. This is likely due
to branch prediction effects out-weighing dirtying the cache for
this particular workload. Also, it is likely that the cache lines are
already dirty as we ray trace neighboring points.