switch from quadratic to gompertz curve expiry
Switch expiration from a quadratic function of "hits" to a gompertz
curve function. The gompertz curve is a good choice for expiration on
hits as it is a non-symmetric, yet easily parameterized signmoid
curve. This allow more control over the expiration when a voxel has
more observations and prevents premature "burn-in". Parameterize the
curve by providing the asymptotes at both positive and negative
infinity, and by providing two more points on the curve. This is
easier to understand intuitively than directly setting the b and c
coefficients of the curve.

Because the octomap server is only used internally, it is OK to break
the ROS parameter interface and no longer accept quadratic
expiration, as the users of the octomap server will be changed
simultaneously to use the new gompertz curve parameterizations.