add opt. sensor_model expected_rate and per_second
Add optional parameters to the sensor_model, expected_rate,
hit_per_second and miss_per_second. If the expected_rate is set, it is
used when calculating the probability of a single hit or miss from the
new hit_per_second and miss_per_second parameters. This makes it easier
to parameterize by time and then to simply set the expected sensor
update rate. Add a throttled warning if the sensor is not updating near
the expected rate.